Moves legend SVG generator into chromosomeSvg.js
It is a chromosome, after all.
